fix(telegram): keep setup single-bot
