+++ docs/troubleshooting.md
@@ -275,10 +275,9 @@ Then check:
 |---|---|
 | Bot never replies | Gateway is not running, the channel is not enabled, or the bot/app token is wrong. |
 | Unknown sender ignored | Configure `allowFrom`, pairing, or the channel-specific allow list. |
-| Telegram support is not installed | In **Settings → Channels**, open Telegram, turn on its switch, and confirm **Install and enable**. If WebUI installation is unavailable, run `nanobot plugins enable telegram` in the gateway's Python environment. |
-| Telegram token is rejected | Copy the current token from BotFather or regenerate it. Do not reuse one token for two bot entries. |
-| Telegram token cannot be checked right now | In **Settings → Channels → Telegram**, open the bot, expand **Advanced**, enter the proxy URL under **Network proxy**, and choose **Check and save**. Or confirm the gateway can reach `api.telegram.org` directly. |
-| A second Telegram webhook bot will not start | Give each enabled webhook bot a different `webhookListenPort`. |
+| Telegram shows a saved configuration but cannot complete a live check | The token is saved. Confirm the gateway can reach `api.telegram.org`, or open **Settings → Channels → Telegram → Advanced → Network proxy** and enter an HTTP or SOCKS proxy. |
+| Telegram rejects the token | Copy the current token from BotFather or regenerate it. |
+| Telegram receives no messages | Confirm the channel is enabled, the gateway is running, and the sender is paired or listed in `allowFrom`. |
 | Discord replies missing | Enable Message Content intent and invite the bot with the required permissions. |
 | WhatsApp or WeChat login expired | Re-run `nanobot channels login whatsapp` or `nanobot channels login weixin`. |
 | Chat app works but WebUI does not | The provider and gateway are likely fine; debug the WebSocket channel separately. |
