--- nanobot/channels/telegram/config.py
@@ -1,73 +1,0 @@
-"""Dependency-free Telegram configuration shared by management and runtime."""
-
-import re
-from typing import Literal
-from urllib.parse import urlparse
-
-from pydantic import Field, field_validator, model_validator
-
-from nanobot.config_base import Base
-
-_STREAM_EDIT_INTERVAL_DEFAULT = 0.6
-
-
-class TelegramConfig(Base):
-    """Configuration for one Telegram bot instance."""
-
-    instance_id: str = "default"
-    name: str = "nanobot"
-    enabled: bool = False
-    token: str = ""
-    mode: Literal["polling", "webhook"] = "polling"
-    allow_from: list[str] = Field(default_factory=list)
-    proxy: str | None = None
-    reply_to_message: bool = False
-    react_emoji: str = "👀"
-    group_policy: Literal["open", "mention"] = "mention"
-    connection_pool_size: int = 32
-    pool_timeout: float = 5.0
-    streaming: bool = True
-    inline_keyboards: bool = False
-    rich_messages: bool = False
-    stream_edit_interval: float = Field(default=_STREAM_EDIT_INTERVAL_DEFAULT, ge=0.1)
-    webhook_url: str = ""
-    webhook_listen_host: str = "127.0.0.1"
-    webhook_listen_port: int = Field(default=8081, ge=1, le=65535)
-    webhook_path: str = "/telegram"
-    webhook_secret_token: str = ""
-    webhook_max_connections: int = Field(default=4, ge=1, le=100)
-
-    @field_validator("webhook_path")
-    @classmethod
-    def webhook_path_must_start_with_slash(cls, value: str) -> str:
-        value = value.strip() or "/telegram"
-        if not value.startswith("/"):
-            raise ValueError('webhook_path must start with "/"')
-        return value
-
-    @model_validator(mode="after")
-    def validate_webhook_config(self) -> "TelegramConfig":
-        if self.mode != "webhook":
-            return self
-
-        url = self.webhook_url.strip()
-        if not url:
-            raise ValueError("webhook_url is required when Telegram mode is webhook")
-        parsed = urlparse(url)
-        if parsed.scheme != "https" or not parsed.netloc:
-            raise ValueError("webhook_url must be a public HTTPS URL")
-        secret = self.webhook_secret_token.strip()
-        if not secret:
-            raise ValueError("webhook_secret_token is required when Telegram mode is webhook")
-        if len(secret) > 256 or re.match(r"^[A-Za-z0-9_-]+$", secret) is None:
-            raise ValueError(
-                "webhook_secret_token must be 1-256 characters using only A-Z, a-z, 0-9, _ and -"
-            )
-        return self
-
-
-def telegram_default_config() -> dict[str, object]:
-    return TelegramConfig().model_dump(by_alias=True)
-
-
-__all__ = ["TelegramConfig", "telegram_default_config"]
