+++ nanobot/channels/telegram/manifest.py
@@ -1,18 +1,16 @@
 """Telegram management contract."""
 
-from nanobot.channels._manifest import DIRECT_GROUP_POLICIES, field, required
+from nanobot.channels._manifest import GROUP_POLICIES, field, required
 from nanobot.channels.contracts import ChannelSetupSpec
 from nanobot.channels.plugin import ChannelPlugin
-from nanobot.channels.telegram.instances import TELEGRAM_MANAGEMENT
 from nanobot.channels.telegram.validation import validate
 
 SETUP_SPEC = ChannelSetupSpec(
     fields={
-        "name": field(),
         "token": field("secret"),
         "proxy": field("secret"),
         "allowFrom": field("list"),
-        "groupPolicy": field("enum", choices=DIRECT_GROUP_POLICIES, default="mention"),
+        "groupPolicy": field("enum", choices=GROUP_POLICIES, default="mention"),
     },
     required=(required("token"),),
     official_url="https://t.me/BotFather",
@@ -24,7 +22,6 @@ PLUGIN = ChannelPlugin(
     display_name="Telegram",
     runtime=f"{__package__}.runtime:TelegramChannel",
     setup=SETUP_SPEC,
-    management=TELEGRAM_MANAGEMENT,
     dependencies=(
         "python-telegram-bot[socks,webhooks]>=22.6,<23.0",
         "socksio>=1.0.0,<2.0.0",
