+++ nanobot/channels/telegram/webui/locales/ko.json
@@ -1,22 +1,21 @@
 {
   "description": "Telegram 채팅에서 nanobot과 대화합니다.",
-  "requirements": "BotFather 토큰과, 네트워크에서 Telegram에 연결할 수 없는 경우 선택적 프록시.",
+  "requirements": "봇 토큰, 허용된 사용자 및 게이트웨이",
   "setup": {
     "docsLabel": "Telegram 설정 가이드 열기",
     "officialLabel": "BotFather 열기",
+    "tryIt": "Telegram 봇에 /start 또는 짧은 DM을 보내세요.",
+    "summary": "활성화하면 Telegram 지원이 켜집니다. 메시지를 주고받으려면 BotFather 토큰이 필요합니다.",
     "steps": [
-      "BotFather에서 봇을 만들고 토큰을 복사합니다.",
-      "여기에 토큰을 붙여 넣습니다. 네트워크에서 Telegram이 차단된 경우 고급 옵션에서 프록시를 추가합니다.",
-      "봇을 연결한 다음 Telegram에서 메시지를 보냅니다."
+      "BotFather로 봇을 만들고 토큰을 복사하세요.",
+      "토큰을 추가하고 봇에 메시지를 보낼 사용자를 선택하세요.",
+      "저장하고 Telegram을 활성화한 다음 DM을 보내거나 그룹에서 멘션하세요."
     ],
     "fields": {
-      "name": {
-        "label": "봇 이름",
-        "placeholder": "지원 봇"
-      },
       "token": {
         "label": "봇 토큰",
-        "placeholder": "123456:ABC..."
+        "placeholder": "123456:ABC...",
+        "help": "BotFather에서 생성하세요."
       },
       "proxy": {
         "label": "네트워크 프록시",
@@ -24,56 +23,17 @@
       },
       "allowFrom": {
         "label": "허용된 사용자",
-        "placeholder": "* 또는 Telegram 사용자 ID"
+        "placeholder": "* 또는 Telegram 사용자 ID",
+        "help": "페어링 코드를 사용하려면 비워 두세요."
       },
       "groupPolicy": {
         "label": "그룹 동작",
         "choices": {
           "mention": "멘션만",
-          "open": "모든 메시지"
+          "open": "모든 메시지",
+          "allowlist": "허용 목록"
         }
       }
     }
-  },
-  "panel": {
-    "toggleBot": "{{name}} 봇",
-    "running": "실행 중",
-    "needsToken": "토큰 필요",
-    "finishSetup": "확인하고 설정 완료",
-    "connectFirst": "첫 번째 봇 연결",
-    "addBot": "봇 추가",
-    "connectBot": "확인하고 연결",
-    "advancedOptions": "고급 옵션",
-    "tokenRequired": "먼저 BotFather 토큰을 입력하세요.",
-    "botName": "봇 이름",
-    "botNamePlaceholder": "지원 봇",
-    "botToken": "봇 토큰",
-    "hideToken": "토큰 숨기기",
-    "showToken": "토큰 표시",
-    "networkProxy": "네트워크 프록시(선택 사항)",
-    "savedProxyPlaceholder": "프록시가 저장되어 있습니다",
-    "hideProxy": "프록시 숨기기",
-    "showProxy": "프록시 표시",
-    "proxyUnavailable": "이 프록시를 통해 Telegram에 연결할 수 없습니다. 주소와 자격 증명을 확인하세요.",
-    "proxyRequired": "먼저 프록시 URL을 입력하세요.",
-    "proxySaved": "프록시를 저장하고 연결을 확인했습니다.",
-    "proxyRemoved": "저장된 프록시를 삭제했습니다.",
-    "proxyConfigured": "설정됨",
-    "saveProxy": "확인 후 저장",
-    "removeProxy": "저장된 프록시 삭제",
-    "invalidProxy": "http://127.0.0.1:7890 같은 전체 프록시 URL을 입력하세요.",
-    "proxyEnvMissing": "저장된 프록시가 참조하는 환경 변수가 설정되지 않았습니다.",
-    "tokenEnvMissing": "저장된 봇 토큰이 참조하는 환경 변수가 설정되지 않았습니다.",
-    "cancel": "취소",
-    "checkConnection": "연결 확인",
-    "connectedAs": "{{name}}(으)로 연결되었습니다.",
-    "connected": "Telegram이 저장된 토큰을 승인했습니다.",
-    "configuredUnverified": "저장된 토큰을 찾았지만 Telegram에 연결할 수 없습니다. 네트워크 또는 프록시를 확인한 후 다시 시도하세요.",
-    "verificationUnavailable": "Telegram에 연결할 수 없습니다. 네트워크를 확인하거나 고급 옵션에서 프록시를 추가한 후 다시 시도하세요.",
-    "invalidToken": "Telegram이 이 봇 토큰을 거부했습니다.",
-    "checkFailed": "연결을 확인할 수 없습니다.",
-    "defaultBotName": "Telegram 봇",
-    "numberedBotName": "Telegram 봇 {{number}}",
-    "count": "설정됨: {{count}} · 실행 중: {{running}}"
   }
 }
