+++ nanobot/channels/telegram/webui/locales/pt-BR.json
@@ -1,22 +1,21 @@
 {
   "description": "Converse com o nanobot pelo Telegram.",
-  "requirements": "Um token do BotFather e, se a sua rede não alcançar o Telegram, um proxy opcional.",
+  "requirements": "Token do bot, usuários permitidos e gateway",
   "setup": {
     "docsLabel": "Abrir guia do Telegram",
     "officialLabel": "Abrir BotFather",
+    "tryIt": "Envie /start ou uma DM curta ao seu bot do Telegram.",
+    "summary": "Ativar liga o suporte ao Telegram. Um token do BotFather ainda é necessário para trocar mensagens.",
     "steps": [
       "Crie um bot com o BotFather e copie o token.",
-      "Cole o token aqui. Se o Telegram estiver bloqueado na sua rede, adicione um proxy nas Opções avançadas.",
-      "Conecte o bot e envie uma mensagem para ele no Telegram."
+      "Adicione o token e escolha quem pode enviar mensagens ao bot.",
+      "Salve e ative o Telegram; depois, envie uma DM ou mencione o bot em um grupo."
     ],
     "fields": {
-      "name": {
-        "label": "Nome do bot",
-        "placeholder": "Bot de suporte"
-      },
       "token": {
         "label": "Token do bot",
-        "placeholder": "123456:ABC..."
+        "placeholder": "123456:ABC...",
+        "help": "Crie-o com o BotFather."
       },
       "proxy": {
         "label": "Proxy de rede",
@@ -24,56 +23,17 @@
       },
       "allowFrom": {
         "label": "Usuários permitidos",
-        "placeholder": "* ou IDs de usuário do Telegram"
+        "placeholder": "* ou IDs de usuário do Telegram",
+        "help": "Deixe vazio para usar códigos de pareamento."
       },
       "groupPolicy": {
         "label": "Comportamento em grupos",
         "choices": {
           "mention": "Somente menções",
-          "open": "Todas as mensagens"
+          "open": "Todas as mensagens",
+          "allowlist": "Lista de permissão"
         }
       }
     }
-  },
-  "panel": {
-    "toggleBot": "Bot {{name}}",
-    "running": "Em execução",
-    "needsToken": "Precisa de token",
-    "finishSetup": "Verificar e concluir",
-    "connectFirst": "Conecte seu primeiro bot",
-    "addBot": "Adicionar bot",
-    "connectBot": "Verificar e conectar",
-    "advancedOptions": "Opções avançadas",
-    "tokenRequired": "Insira primeiro o token do BotFather.",
-    "botName": "Nome do bot",
-    "botNamePlaceholder": "Bot de suporte",
-    "botToken": "Token do bot",
-    "hideToken": "Ocultar token",
-    "showToken": "Mostrar token",
-    "networkProxy": "Proxy de rede (opcional)",
-    "savedProxyPlaceholder": "Já existe um proxy salvo",
-    "hideProxy": "Ocultar proxy",
-    "showProxy": "Mostrar proxy",
-    "proxyUnavailable": "Não foi possível acessar o Telegram por este proxy. Verifique o endereço e as credenciais.",
-    "proxyRequired": "Insira primeiro uma URL de proxy.",
-    "proxySaved": "Proxy salvo e conexão verificada.",
-    "proxyRemoved": "Proxy salvo removido.",
-    "proxyConfigured": "Configurado",
-    "saveProxy": "Verificar e salvar",
-    "removeProxy": "Remover proxy salvo",
-    "invalidProxy": "Insira uma URL de proxy completa, como http://127.0.0.1:7890.",
-    "proxyEnvMissing": "O proxy salvo usa uma variável de ambiente que não está definida.",
-    "tokenEnvMissing": "O token do bot salvo usa uma variável de ambiente que não está definida.",
-    "cancel": "Cancelar",
-    "checkConnection": "Verificar conexão",
-    "connectedAs": "Conectado como {{name}}.",
-    "connected": "O Telegram aceitou o token salvo.",
-    "configuredUnverified": "Um token salvo foi encontrado, mas não foi possível acessar o Telegram. Verifique a rede ou o proxy e tente novamente.",
-    "verificationUnavailable": "Não foi possível acessar o Telegram. Verifique a rede ou adicione um proxy nas Opções avançadas e tente novamente.",
-    "invalidToken": "O Telegram rejeitou este token de bot.",
-    "checkFailed": "Não foi possível verificar a conexão.",
-    "defaultBotName": "Bot do Telegram",
-    "numberedBotName": "Bot do Telegram {{number}}",
-    "count": "Configurados: {{count}} · Em execução: {{running}}"
   }
 }
